third_party/blink/web_tests/external/wpt/clipboard-apis/clipboard-item.https-expected.txt

This is a testharness.js-based test.
[FAIL] getType(DOMString valid type) converts DOMString to Blob
  promise_test: Unhandled rejection with value: object "TypeError: Failed to construct 'ClipboardItem': Failed to convert value to 'Blob'."
[FAIL] getType(DOMString invalid type) converts DOMString to Blob
  promise_test: Unhandled rejection with value: object "TypeError: Failed to construct 'ClipboardItem': Failed to convert value to 'Blob'."
[FAIL] supports(text/uri-list) returns true
  assert_equals: expected true but got false
Harness: the test ran to completion.
